Change config structure
